+++ example/lib/main.dart
@@ -79,13 +79,13 @@ class DemoPod extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
-      title: 'Solid Pod Demonstrator',
+    return MaterialApp(
+      title: AppConstants.longName,
       home: SolidLogin(
         // Images generated using Bing Image Creator from Designer, powered by
         // DALL-E3.
 
-        title: 'SOLID POD DEMONSTRATOR',
+        title: AppConstants.longName.toUpperCase(),
         appDirectory: 'exampleApp',
         image: AssetImage('assets/images/demopod_image.png'),
         logo: AssetImage('assets/images/demopod_logo.png'),